src/main/java/org/example/springproject/config/WebInitializer.java
package org.example.springproject.config;

import org.jetbrains.annotations.NotNull;
import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;

public class WebInitializer extends AbstractAnnotationConfigDispatcherServletInitializer {

    @Override
    protected Class<?>[] getRootConfigClasses() {
        return new Class<?>[]{DatasourceConfig.class, ValidationConfig.class, PropertiesConfig.class};
    }

    @Override
    protected Class<?>[] getServletConfigClasses() {
        return new Class<?>[]{SpringConfig.class};
    }

    @Override
    protected String @NotNull [] getServletMappings() {
        return new String[]{"/*"};
    }
}
